Factor audit code into fn, and add envelope for responses
